pyproject.toml
[project]
name = "miami"
version = "0.1.0"
description = ""
authors = [
    {name = "Razex-ui",email = "ponomarenko_daniil@inbox.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.129.0,<0.130.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "pydantic-extra-types (>=2.11.0,<3.0.0)",
    "phonenumbers (>=9.0.25,<10.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "flake8-pyproject (>=1.2.4,<2.0.0)",
    "psycopg (>=3.3.3,<4.0.0)",
    "psycopg-binary (>=3.3.3,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
test = [
    "pep8-naming (>=0.15.1,<0.16.0)"
]

[tool.flake8]
max-line-length = 120
ignore = [
    # W503: line break before binary operator
    "W503"
]
exclude = [
    ".git",
    "__pycache__",
    "conf",
    "venv",
    ".venv",
    ".vscode",
    ".idea",
    "env.py"
]
per-file-ignores = [
#    "migrations/env.py:F401",
#    "migrations/versions/*:W291",
]